Refactor jhelm-kube into sub-packages
Move HelmKubeService, KubernetesClientProvider → kube.service
Move JhelmKubernetesProperties → kube.config
Keep JhelmKubeAutoConfiguration, KubernetesConfig at root kube package
